requirements.txt
atari-py==0.2.9
certifi==2022.5.18.1
charset-normalizer==2.0.12
cycler==0.11.0
einops==0.8.0
fonttools==4.33.3
idna==3.3
kiwisolver==1.4.2
matplotlib==3.5.2
numpy==1.22.4
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.1.1
plotly==5.10.0
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.27.1
scipy==1.10.1
six==1.16.0
tenacity==8.1.0
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.64.1
typing_extensions==4.2.0
urllib3==1.26.9